backend/requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.7

# Configuration
django-environ==0.4.1
whitenoise==3.3.0


# Forms
django-braces==1.11.0
django-crispy-forms==1.6.1

# Models
django-model-utils==3.0.0

# Images
Pillow==4.1.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.31.0


# Python-PostgreSQL Database Adapter
psycopg2==2.7.1

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2017.2

# Redis support
django-redis==4.7.0
redis>=2.10.5

rcssmin==1.0.6
django-compressor==2.1.1

unicode-slugify==0.1.3

markdown2==2.3.3

django-taggit==0.22.0

# django-secretballot==0.6.0

djangorestframework==3.6.2
django-taggit-serializer==0.1.5

django-rest-auth==0.9.1

django-hitcount==1.2.2

django-webpack-loader==0.6.0
